+++ CHANGES.md
@@ -1,6 +1,7 @@
 [unreleased]
 * add filter `gu_parse_api_branches`
 * add filter `gu_display_repos`
+* add filter `gu_running_git_servers`
 
 #### 0.7.0 / 2021-03-15 🎂
 * add filter `gu_post_api_response_body`